torii.platform.resources: Re-organized platform resources
