+++ torii/platform/resources/user.py
@@ -38,6 +38,16 @@ def _SplitResources(
 
 	return resources
 
+def ButtonResources(
+	name_or_number: str | int | None = None, number: int | None = None, *,
+	pins: str | list[str] | dict[int, str], invert: bool = False,
+	conn: ResourceConn | None = None, attrs: Attrs | None = None,
+) -> list[Resource]:
+	return _SplitResources(
+		name_or_number, number, default_name = 'button', dir = 'i', pins = pins, invert = invert,
+		conn = conn, attrs = attrs
+	)
+
 def LEDResources(
 	name_or_number: str | int | None = None, number: int | None = None, *,
 	pins: str | list[str] | dict[int, str], invert: bool = False,
@@ -63,16 +73,6 @@ def RGBLEDResource(
 		ios.append(attrs)
 
 	return Resource.family(name_or_number, number, default_name = 'rgb_led', ios = ios)
-
-def ButtonResources(
-	name_or_number: str | int | None = None, number: int | None = None, *,
-	pins: str | list[str] | dict[int, str], invert: bool = False,
-	conn: ResourceConn | None = None, attrs: Attrs | None = None,
-) -> list[Resource]:
-	return _SplitResources(
-		name_or_number, number, default_name = 'button', dir = 'i', pins = pins, invert = invert,
-		conn = conn, attrs = attrs
-	)
 
 def SwitchResources(
 	name_or_number: str | int | None = None, number: int | None = None, *,
